Got rid of Windows CRLF line endings.
All non-binary files should be stored in the repository with Unix LF line endings so that git can correctly convert them to Windows CRLF line endings when the source is checked out on a Windows box. Otherwise git is unable to handle CRLF files when core.autocrlf is set to 'true' (recommended on Windows) or 'input' (recommended on Mac/Linux). If you really need a certain file to have CRLF line endings regardless of the platform it is checked out on, unset the crlf attribute on it to make sure git doesn't try to do line ending conversion on it.